+++ packages/go_router_builder/example/lib/main.g.dart
@@ -44,6 +44,8 @@ extension $HomeRouteExtension on HomeRoute {
       );
 
   void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
 }
 
 extension $FamilyRouteExtension on FamilyRoute {
@@ -56,6 +58,8 @@ extension $FamilyRouteExtension on FamilyRoute {
       );
 
   void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
 }
 
 extension $PersonRouteExtension on PersonRoute {
@@ -69,6 +73,8 @@ extension $PersonRouteExtension on PersonRoute {
       );
 
   void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
 }
 
 extension $PersonDetailsRouteExtension on PersonDetailsRoute {
@@ -85,6 +91,8 @@ extension $PersonDetailsRouteExtension on PersonDetailsRoute {
       );
 
   void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
 }
 
 const _$PersonDetailsEnumMap = {
@@ -116,4 +124,6 @@ extension $LoginRouteExtension on LoginRoute {
       );
 
   void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
 }