+++ packages/pigeon/lib/cpp_generator.dart
@@ -92,12 +92,12 @@ void _writeCodecSource(Indent indent, Api api, Root root) {
   indent.writeln('$codecName::$codecName() {}');
   if (getCodecClasses(api, root).isNotEmpty) {
     indent.write(
-        'flutter::EncodableValue $codecName::ReadValueOfType(uint8_t type, flutter::ByteStreamReader* stream) const');
+        'flutter::EncodableValue $codecName::ReadValueOfType(uint8_t type, flutter::ByteStreamReader* stream) const ');
     indent.scoped('{', '}', () {
       indent.write('switch (type) ');
       indent.scoped('{', '}', () {
         for (final EnumeratedClass customClass in getCodecClasses(api, root)) {
-          indent.write('case ${customClass.enumeration}: ');
+          indent.write('case ${customClass.enumeration}:');
           indent.writeScoped('', '', () {
             indent.writeln(
                 'return flutter::CustomEncodableValue(${customClass.name}(std::get<flutter::EncodableMap>(ReadValue(stream))));');
@@ -107,19 +107,19 @@ void _writeCodecSource(Indent indent, Api api, Root root) {
         indent.writeScoped('', '', () {
           indent.writeln(
               'return flutter::StandardCodecSerializer::ReadValueOfType(type, stream);');
-        });
+        }, addTrailingNewline: false);
       });
     });
     indent.writeln('');
     indent.write(
-        'void $codecName::WriteValue(const flutter::EncodableValue& value, flutter::ByteStreamWriter* stream) const');
+        'void $codecName::WriteValue(const flutter::EncodableValue& value, flutter::ByteStreamWriter* stream) const ');
     indent.writeScoped('{', '}', () {
       indent.write(
-          'if (const flutter::CustomEncodableValue* custom_value = std::get_if<flutter::CustomEncodableValue>(&value))');
+          'if (const flutter::CustomEncodableValue* custom_value = std::get_if<flutter::CustomEncodableValue>(&value)) ');
       indent.scoped('{', '}', () {
         for (final EnumeratedClass customClass in getCodecClasses(api, root)) {
           indent.write(
-              'if (custom_value->type() == typeid(${customClass.name}))');
+              'if (custom_value->type() == typeid(${customClass.name})) ');
           indent.scoped('{', '}', () {
             indent.writeln('stream->WriteByte(${customClass.enumeration});');
             indent.writeln(
@@ -171,6 +171,220 @@ template<class T> class ErrorOr {
 ''');
 }
 
+/// Writes the declaration for the custom class [klass].
+///
+/// See [_writeDataClassImplementation] for the corresponding declaration.
+/// This is intended to be added to the header.
+void _writeDataClassDeclaration(Indent indent, Class klass, Root root,
+    {String? testFriend}) {
+  indent.addln('');
+  indent.writeln(
+      '/* Generated class from Pigeon that represents data sent in messages. */');
+  indent.write('class ${klass.name} ');
+  indent.scoped('{', '};', () {
+    indent.scoped(' public:', '', () {
+      indent.writeln('${klass.name}();');
+      for (final NamedType field in klass.fields) {
+        final HostDatatype baseDatatype = getHostDatatype(
+            field,
+            root.classes,
+            root.enums,
+            (NamedType x) => _baseCppTypeForBuiltinDartType(x.type));
+        indent.writeln(
+            '${_getterReturnType(baseDatatype)} ${_makeGetterName(field)}() const;');
+        indent.writeln(
+            'void ${_makeSetterName(field)}(${_unownedArgumentType(baseDatatype)} value_arg);');
+        if (field.type.isNullable) {
+          // Add a second setter that takes the non-nullable version of the
+          // argument for convenience, since setting literal values with the
+          // pointer version is non-trivial.
+          final HostDatatype nonNullType = _nonNullableType(baseDatatype);
+          indent.writeln(
+              'void ${_makeSetterName(field)}(${_unownedArgumentType(nonNullType)} value_arg);');
+        }
+        indent.addln('');
+      }
+    });
+
+    indent.scoped(' private:', '', () {
+      indent.writeln('${klass.name}(flutter::EncodableMap map);');
+      indent.writeln('flutter::EncodableMap ToEncodableMap() const;');
+      for (final Class friend in root.classes) {
+        if (friend != klass &&
+            friend.fields.any(
+                (NamedType element) => element.type.baseName == klass.name)) {
+          indent.writeln('friend class ${friend.name};');
+        }
+      }
+      for (final Api api in root.apis) {
+        // TODO(gaaclarke): Find a way to be more precise with our
+        // friendships.
+        indent.writeln('friend class ${api.name};');
+        indent.writeln('friend class ${_getCodecName(api)};');
+      }
+      if (testFriend != null) {
+        indent.writeln('friend class $testFriend;');
+      }
+
+      for (final NamedType field in klass.fields) {
+        final HostDatatype hostDatatype = getHostDatatype(
+            field,
+            root.classes,
+            root.enums,
+            (NamedType x) => _baseCppTypeForBuiltinDartType(x.type));
+        indent.writeln(
+            '${_valueType(hostDatatype)} ${_makeInstanceVariableName(field)};');
+      }
+    });
+  }, nestCount: 0);
+  indent.writeln('');
+}
+
+/// Writes the implementation for the custom class [klass].
+///
+/// See [_writeDataClassDeclaration] for the corresponding declaration.
+/// This is intended to be added to the implementation file.
+void _writeDataClassImplementation(Indent indent, Class klass, Root root) {
+  final Set<String> rootClassNameSet =
+      root.classes.map((Class x) => x.name).toSet();
+  final Set<String> rootEnumNameSet =
+      root.enums.map((Enum x) => x.name).toSet();
+
+  indent.addln('');
+  indent.writeln('/* ${klass.name} */');
+  indent.addln('');
+
+  // Getters and setters.
+  for (final NamedType field in klass.fields) {
+    final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
+        root.enums, (NamedType x) => _baseCppTypeForBuiltinDartType(x.type));
+    final String instanceVariableName = _makeInstanceVariableName(field);
+    final String qualifiedGetterName =
+        '${klass.name}::${_makeGetterName(field)}';
+    final String qualifiedSetterName =
+        '${klass.name}::${_makeSetterName(field)}';
+    final String returnExpression = hostDatatype.isNullable
+        ? '$instanceVariableName ? &(*$instanceVariableName) : nullptr'
+        : instanceVariableName;
+
+    // Generates the string for a setter treating the type as [type], to allow
+    // generating multiple setter variants.
+    String makeSetter(HostDatatype type) {
+      const String setterArgumentName = 'value_arg';
+      final String valueExpression = type.isNullable
+          ? '$setterArgumentName ? ${_valueType(type)}(*$setterArgumentName) : std::nullopt'
+          : setterArgumentName;
+      return 'void $qualifiedSetterName(${_unownedArgumentType(type)} $setterArgumentName) '
+          '{ $instanceVariableName = $valueExpression; }';
+    }
+
+    indent.writeln(
+        '${_getterReturnType(hostDatatype)} $qualifiedGetterName() const '
+        '{ return $returnExpression; }');
+    indent.writeln(makeSetter(hostDatatype));
+    if (hostDatatype.isNullable) {
+      // Write the non-nullable variant; see _writeDataClassDeclaration.
+      final HostDatatype nonNullType = _nonNullableType(hostDatatype);
+      indent.writeln(makeSetter(nonNullType));
+    }
+
+    indent.addln('');
+  }
+
+  // Serialization.
+  indent.write('flutter::EncodableMap ${klass.name}::ToEncodableMap() const ');
+  indent.scoped('{', '}', () {
+    indent.scoped('return flutter::EncodableMap{', '};', () {
+      for (final NamedType field in klass.fields) {
+        final HostDatatype hostDatatype = getHostDatatype(
+            field,
+            root.classes,
+            root.enums,
+            (NamedType x) => _baseCppTypeForBuiltinDartType(x.type));
+
+        final String instanceVariable = _makeInstanceVariableName(field);
+
+        final String encodableKey = 'flutter::EncodableValue("${field.name}")';
+        String encodableValue = '';
+        if (!hostDatatype.isBuiltin &&
+            rootClassNameSet.contains(field.type.baseName)) {
+          final String operator = field.type.isNullable ? '->' : '.';
+          encodableValue = '$instanceVariable${operator}ToEncodableMap()';
+        } else if (!hostDatatype.isBuiltin &&
+            rootEnumNameSet.contains(field.type.baseName)) {
+          final String nonNullValue =
+              field.type.isNullable ? '(*$instanceVariable)' : instanceVariable;
+          encodableValue = 'flutter::EncodableValue((int)$nonNullValue)';
+        } else {
+          final String operator = field.type.isNullable ? '*' : '';
+          encodableValue =
+              'flutter::EncodableValue($operator$instanceVariable)';
+        }
+
+        if (field.type.isNullable) {
+          encodableValue =
+              '$instanceVariable ? $encodableValue : flutter::EncodableValue()';
+        }
+
+        indent.writeln('{$encodableKey, $encodableValue},');
+      }
+    });
+  });
+  indent.addln('');
+
+  // Default constructor.
+  indent.writeln('${klass.name}::${klass.name}() {}');
+  indent.addln('');
+
+  // Deserialization.
+  indent.write('${klass.name}::${klass.name}(flutter::EncodableMap map) ');
+  indent.scoped('{', '}', () {
+    for (final NamedType field in klass.fields) {
+      final String instanceVariableName = _makeInstanceVariableName(field);
+      final String pointerFieldName =
+          '${_pointerPrefix}_${_makeVariableName(field)}';
+      final String encodableFieldName =
+          '${_encodablePrefix}_${_makeVariableName(field)}';
+      indent.writeln(
+          'auto $encodableFieldName = map.at(flutter::EncodableValue("${field.name}"));');
+      if (rootEnumNameSet.contains(field.type.baseName)) {
+        indent.writeln(
+            'if (const int32_t* $pointerFieldName = std::get_if<int32_t>(&$encodableFieldName))\t$instanceVariableName = (${field.type.baseName})*$pointerFieldName;');
+      } else {
+        final HostDatatype hostDatatype = getHostDatatype(
+            field,
+            root.classes,
+            root.enums,
+            (NamedType x) => _baseCppTypeForBuiltinDartType(x.type));
+        if (field.type.baseName == 'int') {
+          indent.format('''
+if (const int32_t* $pointerFieldName = std::get_if<int32_t>(&$encodableFieldName))
+\t$instanceVariableName = *$pointerFieldName;
+else if (const int64_t* ${pointerFieldName}_64 = std::get_if<int64_t>(&$encodableFieldName))
+\t$instanceVariableName = *${pointerFieldName}_64;''');
+        } else if (!hostDatatype.isBuiltin &&
+            root.classes
+                .map((Class x) => x.name)
+                .contains(field.type.baseName)) {
+          indent.write(
+              'if (const flutter::EncodableMap* $pointerFieldName = std::get_if<flutter::EncodableMap>(&$encodableFieldName)) ');
+          indent.scoped('{', '}', () {
+            indent.writeln(
+                '$instanceVariableName = ${hostDatatype.datatype}(*$pointerFieldName);');
+          });
+        } else {
+          indent.write(
+              'if (const ${hostDatatype.datatype}* $pointerFieldName = std::get_if<${hostDatatype.datatype}>(&$encodableFieldName)) ');
+          indent.scoped('{', '}', () {
+            indent.writeln('$instanceVariableName = *$pointerFieldName;');
+          });
+        }
+      }
+    }
+  });
+  indent.addln('');
+}
+
 void _writeHostApiHeader(Indent indent, Api api) {
   assert(api.location == ApiLocation.host);
 
@@ -192,7 +406,7 @@ void _writeHostApiHeader(Indent indent, Api api) {
           final Iterable<String> argTypes = method.arguments
               .map((NamedType e) => _nullSafeCppTypeForDartType(e.type));
           final Iterable<String> argNames =
-              method.arguments.map((NamedType e) => e.name);
+              method.arguments.map((NamedType e) => _makeVariableName(e));
           argSignature.addAll(
               map2(argTypes, argNames, (String argType, String argName) {
             return '$argType $argName';
@@ -201,10 +415,10 @@ void _writeHostApiHeader(Indent indent, Api api) {
         if (method.isAsynchronous) {
           argSignature.add('std::function<void($returnTypeName reply)> result');
           indent.writeln(
-              'virtual void ${method.name}(${argSignature.join(', ')}) = 0;');
+              'virtual void ${_makeMethodName(method)}(${argSignature.join(', ')}) = 0;');
         } else {
           indent.writeln(
-              'virtual $returnTypeName ${method.name}(${argSignature.join(', ')}) = 0;');
+              'virtual $returnTypeName ${_makeMethodName(method)}(${argSignature.join(', ')}) = 0;');
         }
       }
       indent.addln('');
@@ -215,7 +429,7 @@ void _writeHostApiHeader(Indent indent, Api api) {
       indent.writeln(
           'static void SetUp(flutter::BinaryMessenger* binary_messenger, ${api.name}* api);');
       indent.writeln(
-          'static flutter::EncodableMap WrapError(std::string_view errorMessage);');
+          'static flutter::EncodableMap WrapError(std::string_view error_message);');
       indent.writeln(
           'static flutter::EncodableMap WrapError(const FlutterError& error);');
     });
@@ -254,7 +468,7 @@ const flutter::StandardMessageCodec& ${api.name}::GetCodec() {
         indent.write('if (api != nullptr) ');
         indent.scoped('{', '} else {', () {
           indent.write(
-              'channel->SetMessageHandler([api](const flutter::EncodableValue& message, const flutter::MessageReply<flutter::EncodableValue>& reply)');
+              'channel->SetMessageHandler([api](const flutter::EncodableValue& message, const flutter::MessageReply<flutter::EncodableValue>& reply) ');
           indent.scoped('{', '});', () {
             final String returnTypeName = method.returnType.isVoid
                 ? 'std::optional<FlutterError>'
@@ -289,8 +503,8 @@ const flutter::StandardMessageCodec& ${api.name}::GetCodec() {
               }
 
               String _wrapResponse(String reply, TypeDeclaration returnType) {
-                final bool isReferenceReturnType =
-                    _isReferenceType(_cppTypeForDartType(method.returnType));
+                final bool isReferenceReturnType = _isReferenceType(
+                    _baseCppTypeForDartType(method.returnType));
                 String elseBody = '';
                 final String ifCondition;
                 final String errorGetter;
@@ -332,7 +546,7 @@ $prefix\t}${indent.newline}''';
                 );
               }
               final String call =
-                  'api->${method.name}(${methodArgument.join(', ')})';
+                  'api->${_makeMethodName(method)}(${methodArgument.join(', ')})';
               if (method.isAsynchronous) {
                 indent.format('$call;');
               } else {
@@ -362,7 +576,7 @@ $prefix\t}${indent.newline}''';
 }
 
 String _getArgumentName(int count, NamedType argument) =>
-    argument.name.isEmpty ? 'arg$count' : argument.name;
+    argument.name.isEmpty ? 'arg$count' : _makeVariableName(argument);
 
 /// Returns an argument name that can be used in a context where it is possible to collide.
 String _getSafeArgumentName(int count, NamedType argument) =>
@@ -408,7 +622,7 @@ void _writeFlutterApiSource(Indent indent, Api api) {
   indent.writeln(
       '/* Generated class from Pigeon that represents Flutter messages that can be called from C++. */');
   indent.write(
-      '${api.name}::${api.name}(flutter::BinaryMessenger* binary_messenger)');
+      '${api.name}::${api.name}(flutter::BinaryMessenger* binary_messenger) ');
   indent.scoped('{', '}', () {
     indent.writeln('this->binary_messenger_ = binary_messenger;');
   });
@@ -451,7 +665,7 @@ const flutter::StandardMessageCodec& ${api.name}::GetCodec() {
       indent.dec();
       indent.dec();
       indent.write(
-          '$channel->Send($sendArgument, [callback](const uint8_t* reply, size_t reply_size)');
+          '$channel->Send($sendArgument, [callback](const uint8_t* reply, size_t reply_size) ');
       indent.scoped('{', '});', () {
         if (func.returnType.isVoid) {
           indent.writeln('callback();');
@@ -463,8 +677,9 @@ const flutter::StandardMessageCodec& ${api.name}::GetCodec() {
           const String output = 'output';
 
           final bool isBuiltin =
-              _cppTypeForBuiltinDartType(func.returnType) != null;
-          final String returnTypeName = _cppTypeForDartType(func.returnType);
+              _baseCppTypeForBuiltinDartType(func.returnType) != null;
+          final String returnTypeName =
+              _baseCppTypeForDartType(func.returnType);
           if (func.returnType.isNullable) {
             indent.writeln('$returnType $output{};');
           } else {
@@ -479,13 +694,13 @@ else if (const int64_t* ${pointerVariable}_64 = std::get_if<int64_t>(&args))
 \t$output = *${pointerVariable}_64;''');
           } else if (!isBuiltin) {
             indent.write(
-                'if (const flutter::EncodableMap* $pointerVariable = std::get_if<flutter::EncodableMap>(&args))');
+                'if (const flutter::EncodableMap* $pointerVariable = std::get_if<flutter::EncodableMap>(&args)) ');
             indent.scoped('{', '}', () {
               indent.writeln('$output = $returnTypeName(*$pointerVariable);');
             });
           } else {
             indent.write(
-                'if (const $returnTypeName* $pointerVariable = std::get_if<$returnTypeName>(&args))');
+                'if (const $returnTypeName* $pointerVariable = std::get_if<$returnTypeName>(&args)) ');
             indent.scoped('{', '}', () {
               indent.writeln('$output = *$pointerVariable;');
             });
@@ -498,18 +713,41 @@ else if (const int64_t* ${pointerVariable}_64 = std::get_if<int64_t>(&args))
   }
 }
 
-String _makeGetterName(NamedType field) {
-  final String uppercased =
-      field.name.substring(0, 1).toUpperCase() + field.name.substring(1);
-  return 'get$uppercased';
+/// Returns a non-nullable variant of [type].
+HostDatatype _nonNullableType(HostDatatype type) {
+  return HostDatatype(
+      datatype: type.datatype, isBuiltin: type.isBuiltin, isNullable: false);
 }
 
-String _makeSetterName(NamedType field) {
-  final String uppercased =
-      field.name.substring(0, 1).toUpperCase() + field.name.substring(1);
-  return 'set$uppercased';
+String _pascalCaseFromCamelCase(String camelCase) =>
+    camelCase[0].toUpperCase() + camelCase.substring(1);
+
+String _snakeCaseFromCamelCase(String camelCase) {
+  return camelCase.replaceAllMapped(RegExp(r'[A-Z]'),
+      (Match m) => '${m.start == 0 ? '' : '_'}${m[0]!.toLowerCase()}');
 }
 
+String _pascalCaseFromSnakeCase(String snakeCase) {
+  final String camelCase = snakeCase.replaceAllMapped(
+      RegExp(r'_([a-z])'), (Match m) => m[1]!.toUpperCase());
+  return _pascalCaseFromCamelCase(camelCase);
+}
+
+String _makeMethodName(Method method) => _pascalCaseFromCamelCase(method.name);
+
+String _makeGetterName(NamedType field) => _snakeCaseFromCamelCase(field.name);
+
+String _makeSetterName(NamedType field) =>
+    'set_${_snakeCaseFromCamelCase(field.name)}';
+
+String _makeVariableName(NamedType field) =>
+    _snakeCaseFromCamelCase(field.name);
+
+String _makeInstanceVariableName(NamedType field) =>
+    '${_makeVariableName(field)}_';
+
+// TODO(stuartmorgan): Remove this in favor of _isPodType once callers have
+// all been updated to using HostDatatypes.
 bool _isReferenceType(String dataType) {
   switch (dataType) {
     case 'bool':
@@ -521,7 +759,13 @@ bool _isReferenceType(String dataType) {
   }
 }
 
-String? _cppTypeForBuiltinDartType(TypeDeclaration type) {
+/// Returns true if [type] corresponds to a plain-old-data type (i.e., one that
+/// should generally be passed by value rather than pointer/reference) in C++.
+bool _isPodType(HostDatatype type) {
+  return !_isReferenceType(type.datatype);
+}
+
+String? _baseCppTypeForBuiltinDartType(TypeDeclaration type) {
   const Map<String, String> cppTypeForDartTypeMap = <String, String>{
     'bool': 'bool',
     'int': 'int64_t',
@@ -541,16 +785,51 @@ String? _cppTypeForBuiltinDartType(TypeDeclaration type) {
   }
 }
 
-String _cppTypeForDartType(TypeDeclaration type) {
-  return _cppTypeForBuiltinDartType(type) ?? type.baseName;
+/// Returns the base C++ type (without pointer, reference, optional, etc.) for
+/// the given [type].
+String _baseCppTypeForDartType(TypeDeclaration type) {
+  return _baseCppTypeForBuiltinDartType(type) ?? type.baseName;
 }
 
+/// Returns the C++ type to use in a value context (variable declaration,
+/// pass-by-value, etc.) for the given C++ base type.
+String _valueType(HostDatatype type) {
+  final String baseType = type.datatype;
+  return type.isNullable ? 'std::optional<$baseType>' : baseType;
+}
+
+/// Returns the C++ type to use in an argument context without ownership
+/// transfer for the given base type.
+String _unownedArgumentType(HostDatatype type) {
+  final bool isString = type.datatype == 'std::string';
+  final String baseType = isString ? 'std::string_view' : type.datatype;
+  if (isString || _isPodType(type)) {
+    return type.isNullable ? 'const $baseType*' : baseType;
+  }
+  return type.isNullable ? 'const $baseType*' : 'const $baseType&';
+}
+
+/// Returns the C++ type to use for the return of a getter for a field of type
+/// [type].
+String _getterReturnType(HostDatatype type) {
+  final String baseType = type.datatype;
+  if (_isPodType(type)) {
+    // Use pointers rather than optionals even for nullable POD, since the
+    // semantics of using them is essentially identical and this makes them
+    // consistent with non-POD.
+    return type.isNullable ? 'const $baseType*' : baseType;
+  }
+  return type.isNullable ? 'const $baseType*' : 'const $baseType&';
+}
+
+// TODO(stuartmorgan): Audit all uses of this and convert them to context-based
+// methods like those above. Code still using this method may well have bugs.
 String _nullSafeCppTypeForDartType(TypeDeclaration type,
     {bool considerReference = true}) {
   if (type.isNullable) {
-    return 'std::optional<${_cppTypeForDartType(type)}>';
+    return 'std::optional<${_baseCppTypeForDartType(type)}>';
   } else {
-    String typeName = _cppTypeForDartType(type);
+    String typeName = _baseCppTypeForDartType(type);
     if (_isReferenceType(typeName)) {
       if (considerReference)
         typeName = 'const $typeName&';
@@ -572,6 +851,13 @@ String _getGuardName(String? headerFileName, String? namespace) {
   return guardName + 'H_';
 }
 
+void _writeSystemHeaderIncludeBlock(Indent indent, List<String> headers) {
+  headers.sort();
+  for (final String header in headers) {
+    indent.writeln('#include <$header>');
+  }
+}
+
 /// Generates the ".h" file for the AST represented by [root] to [sink] with the
 /// provided [options] and [headerFileName].
 void generateCppHeader(
@@ -586,19 +872,33 @@ void generateCppHeader(
   final String guardName = _getGuardName(headerFileName, options.namespace);
   indent.writeln('#ifndef $guardName');
   indent.writeln('#define $guardName');
-  indent.writeln('#include <flutter/encodable_value.h>');
-  indent.writeln('#include <flutter/basic_message_channel.h>');
-  indent.writeln('#include <flutter/binary_messenger.h>');
-  indent.writeln('#include <flutter/standard_message_codec.h>');
-  indent.addln('');
-  indent.writeln('#include <map>');
-  indent.writeln('#include <string>');
-  indent.writeln('#include <optional>');
 
+  _writeSystemHeaderIncludeBlock(indent, <String>[
+    'flutter/basic_message_channel.h',
+    'flutter/binary_messenger.h',
+    'flutter/encodable_value.h',
+    'flutter/standard_message_codec.h',
+  ]);
+  indent.addln('');
+  _writeSystemHeaderIncludeBlock(indent, <String>[
+    'map',
+    'string',
+    'optional',
+  ]);
   indent.addln('');
 
   if (options.namespace != null) {
     indent.writeln('namespace ${options.namespace} {');
+  }
+
+  // When generating for a Pigeon unit test, add a test fixture friend class to
+  // allow unit testing private methods, since testing serialization via public
+  // methods is essentially an end-to-end test.
+  String? testFixtureClass;
+  if (options.namespace?.endsWith('_pigeontest') ?? false) {
+    testFixtureClass =
+        '${_pascalCaseFromSnakeCase(options.namespace!.replaceAll('_pigeontest', ''))}Test';
+    indent.writeln('class $testFixtureClass;');
   }
 
   indent.addln('');
@@ -622,56 +922,10 @@ void generateCppHeader(
   _writeErrorOr(indent);
 
   for (final Class klass in root.classes) {
-    indent.addln('');
-    indent.writeln(
-        '/* Generated class from Pigeon that represents data sent in messages. */');
-    indent.write('class ${klass.name} ');
-    indent.scoped('{', '};', () {
-      indent.scoped(' public:', '', () {
-        indent.writeln('${klass.name}();');
-        for (final NamedType field in klass.fields) {
-          final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
-              root.enums, (NamedType x) => _cppTypeForBuiltinDartType(x.type));
-          if (_isReferenceType(hostDatatype.datatype)) {
-            indent.writeln(
-                'const ${hostDatatype.datatype}& ${_makeGetterName(field)}() const;');
-            indent.writeln(
-                'void ${_makeSetterName(field)}(const ${hostDatatype.datatype}& setterArg);');
-          } else {
-            indent.writeln(
-                '${hostDatatype.datatype} ${_makeGetterName(field)}() const;');
-            indent.writeln(
-                'void ${_makeSetterName(field)}(const ${hostDatatype.datatype} setterArg);');
-          }
-          indent.addln('');
-        }
-      });
-
-      indent.scoped(' private:', '', () {
-        indent.writeln('${klass.name}(flutter::EncodableMap map);');
-        indent.writeln('flutter::EncodableMap ToEncodableMap();');
-        for (final Class friend in root.classes) {
-          if (friend != klass &&
-              friend.fields.any(
-                  (NamedType element) => element.type.baseName == klass.name)) {
-            indent.writeln('friend class ${friend.name};');
-          }
-        }
-        for (final Api api in root.apis) {
-          // TODO(gaaclarke): Find a way to be more precise with our
-          // friendships.
-          indent.writeln('friend class ${api.name};');
-          indent.writeln('friend class ${_getCodecName(api)};');
-        }
-
-        for (final NamedType field in klass.fields) {
-          final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
-              root.enums, (NamedType x) => _cppTypeForBuiltinDartType(x.type));
-          indent.writeln('${hostDatatype.datatype} ${field.name}_;');
-        }
-      });
-    }, nestCount: 0);
-    indent.writeln('');
+    _writeDataClassDeclaration(indent, klass, root,
+        // Add a hook for unit testing data classes when using the namespace
+        // used by pigeon tests.
+        testFriend: testFixtureClass);
   }
 
   for (final Api api in root.apis) {
@@ -685,7 +939,7 @@ void generateCppHeader(
   }
 
   if (options.namespace != null) {
-    indent.writeln('} // namespace');
+    indent.writeln('}  // namespace ${options.namespace}');
   }
 
   indent.writeln('#endif  // $guardName');
@@ -694,10 +948,6 @@ void generateCppHeader(
 /// Generates the ".cpp" file for the AST represented by [root] to [sink] with the
 /// provided [options].
 void generateCppSource(CppOptions options, Root root, StringSink sink) {
-  final Set<String> rootClassNameSet =
-      root.classes.map((Class x) => x.name).toSet();
-  final Set<String> rootEnumNameSet =
-      root.enums.map((Enum x) => x.name).toSet();
   final Indent indent = Indent(sink);
   if (options.copyrightHeader != null) {
     addLines(indent, options.copyrightHeader!, linePrefix: '// ');
@@ -707,109 +957,29 @@ void generateCppSource(CppOptions options, Root root, StringSink sink) {
   indent.addln('');
   indent.addln('#undef _HAS_EXCEPTIONS');
   indent.addln('');
-  indent.writeln('#include <flutter/basic_message_channel.h>');
-  indent.writeln('#include <flutter/binary_messenger.h>');
-  indent.writeln('#include <flutter/standard_message_codec.h>');
-  indent.writeln('#include <map>');
-  indent.writeln('#include <string>');
-  indent.writeln('#include <optional>');
 
   indent.writeln('#include "${options.header}"');
-
   indent.addln('');
-
+  _writeSystemHeaderIncludeBlock(indent, <String>[
+    'flutter/basic_message_channel.h',
+    'flutter/binary_messenger.h',
+    'flutter/encodable_value.h',
+    'flutter/standard_message_codec.h',
+  ]);
+  indent.addln('');
+  _writeSystemHeaderIncludeBlock(indent, <String>[
+    'map',
+    'string',
+    'optional',
+  ]);
   indent.addln('');
 
   if (options.namespace != null) {
     indent.writeln('namespace ${options.namespace} {');
   }
 
-  indent.addln('');
-  indent.writeln('/* Generated class from Pigeon. */');
-
   for (final Class klass in root.classes) {
-    indent.addln('');
-    indent.writeln('/* ${klass.name} */');
-    indent.addln('');
-    for (final NamedType field in klass.fields) {
-      final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
-          root.enums, (NamedType x) => _cppTypeForBuiltinDartType(x.type));
-      if (_isReferenceType(hostDatatype.datatype)) {
-        indent.writeln(
-            'const ${hostDatatype.datatype}& ${klass.name}::${_makeGetterName(field)}() const { return ${field.name}_; }');
-        indent.writeln(
-            'void ${klass.name}::${_makeSetterName(field)}(const ${hostDatatype.datatype}& setterArg) { this->${field.name}_ = setterArg; }');
-      } else {
-        indent.writeln(
-            '${hostDatatype.datatype} ${klass.name}::${_makeGetterName(field)}() const { return ${field.name}_; }');
-        indent.writeln(
-            'void ${klass.name}::${_makeSetterName(field)}(const ${hostDatatype.datatype} setterArg) { this->${field.name}_ = setterArg; }');
-      }
-
-      indent.addln('');
-    }
-    indent.write('flutter::EncodableMap ${klass.name}::ToEncodableMap() ');
-    indent.scoped('{', '}', () {
-      indent.writeln('flutter::EncodableMap toMapResult;');
-      for (final NamedType field in klass.fields) {
-        final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
-            root.enums, (NamedType x) => _cppTypeForBuiltinDartType(x.type));
-        String toWriteValue = '';
-        if (!hostDatatype.isBuiltin &&
-            rootClassNameSet.contains(field.type.baseName)) {
-          toWriteValue = '${field.name}_.ToEncodableMap()';
-        } else if (!hostDatatype.isBuiltin &&
-            rootEnumNameSet.contains(field.type.baseName)) {
-          toWriteValue = 'flutter::EncodableValue((int)${field.name}_)';
-        } else {
-          toWriteValue = 'flutter::EncodableValue(${field.name}_)';
-        }
-        indent.writeln(
-            'toMapResult.insert(std::make_pair(flutter::EncodableValue("${field.name}"), $toWriteValue));');
-      }
-      indent.writeln('return toMapResult;');
-    });
-    indent.writeln('${klass.name}::${klass.name}() {}');
-    indent.write('${klass.name}::${klass.name}(flutter::EncodableMap map) ');
-    indent.scoped('{', '}', () {
-      for (final NamedType field in klass.fields) {
-        final String pointerFieldName = '${_pointerPrefix}_${field.name}';
-        final String encodableFieldName = '${_encodablePrefix}_${field.name}';
-        indent.writeln(
-            'auto $encodableFieldName = map.at(flutter::EncodableValue("${field.name}"));');
-        if (rootEnumNameSet.contains(field.type.baseName)) {
-          indent.writeln(
-              'if (const int32_t* $pointerFieldName = std::get_if<int32_t>(&$encodableFieldName))\t${field.name}_ = (${field.type.baseName})*$pointerFieldName;');
-        } else {
-          final HostDatatype hostDatatype = getHostDatatype(field, root.classes,
-              root.enums, (NamedType x) => _cppTypeForBuiltinDartType(x.type));
-          if (field.type.baseName == 'int') {
-            indent.format('''
-if (const int32_t* $pointerFieldName = std::get_if<int32_t>(&$encodableFieldName))
-\t${field.name}_ = *$pointerFieldName;
-else if (const int64_t* ${pointerFieldName}_64 = std::get_if<int64_t>(&$encodableFieldName))
-\t${field.name}_ = *${pointerFieldName}_64;''');
-          } else if (!hostDatatype.isBuiltin &&
-              root.classes
-                  .map((Class x) => x.name)
-                  .contains(field.type.baseName)) {
-            indent.write(
-                'if (const flutter::EncodableMap* $pointerFieldName = std::get_if<flutter::EncodableMap>(&$encodableFieldName))');
-            indent.scoped('{', '}', () {
-              indent.writeln(
-                  '${field.name}_ = ${hostDatatype.datatype}(*$pointerFieldName);');
-            });
-          } else {
-            indent.write(
-                'if (const ${hostDatatype.datatype}* $pointerFieldName = std::get_if<${hostDatatype.datatype}>(&$encodableFieldName))');
-            indent.scoped('{', '}', () {
-              indent.writeln('${field.name}_ = *$pointerFieldName;');
-            });
-          }
-        }
-      }
-    });
-    indent.addln('');
+    _writeDataClassImplementation(indent, klass, root);
   }
 
   for (final Api api in root.apis) {
@@ -820,9 +990,9 @@ else if (const int64_t* ${pointerFieldName}_64 = std::get_if<int64_t>(&$encodabl
 
       indent.addln('');
       indent.format('''
-flutter::EncodableMap ${api.name}::WrapError(std::string_view errorMessage) {
+flutter::EncodableMap ${api.name}::WrapError(std::string_view error_message) {
 \treturn flutter::EncodableMap({
-\t\t{flutter::EncodableValue("${Keys.errorMessage}"), flutter::EncodableValue(std::string(errorMessage).data())},
+\t\t{flutter::EncodableValue("${Keys.errorMessage}"), flutter::EncodableValue(std::string(error_message).data())},
 \t\t{flutter::EncodableValue("${Keys.errorCode}"), flutter::EncodableValue("Error")},
 \t\t{flutter::EncodableValue("${Keys.errorDetails}"), flutter::EncodableValue()}
 \t});
@@ -841,20 +1011,22 @@ flutter::EncodableMap ${api.name}::WrapError(const FlutterError& error) {
   }
 
   if (options.namespace != null) {
-    indent.writeln('} // namespace');
+    indent.writeln('}  // namespace ${options.namespace}');
   }
 }
 
 /// Validates an AST to make sure the cpp generator supports everything.
 List<Error> validateCpp(CppOptions options, Root root) {
   final List<Error> result = <Error>[];
-  for (final Class aClass in root.classes) {
-    for (final NamedType field in aClass.fields) {
-      if (!field.type.isNullable) {
-        // TODO(gaaclarke): Add line number and filename.
-        result.add(Error(
-            message:
-                'unsupported nonnull field "${field.name}" in "${aClass.name}"'));
+  for (final Api api in root.apis) {
+    for (final Method method in api.methods) {
+      for (final NamedType arg in method.arguments) {
+        if (isEnum(root, arg.type)) {
+          // TODO(gaaclarke): Add line number and filename.
+          result.add(Error(
+              message:
+                  'Nullable enum types aren\'t supported in C++ arguments in method:${api.name}.${method.name} argument:(${arg.type.baseName} ${arg.name}).'));
+        }
       }
     }
   }