+++ packages/rfw/example/wasm/lib/main.dart
@@ -14,7 +14,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:rfw/rfw.dart';
 import 'package:wasm/wasm.dart';
 
-const String urlPrefix = 'https://raw.githubusercontent.com/flutter/packages/master/packages/rfw/example/wasm/logic';
+const String urlPrefix = 'https://raw.githubusercontent.com/flutter/packages/main/packages/rfw/example/wasm/logic';
 
 const String interfaceUrl = '$urlPrefix/calculator.rfw';
 const String logicUrl = '$urlPrefix/calculator.wasm';