+++ packages/flutter_image/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.1.3
+
+* Suppresses deprecation warnings.
+
+## 4.1.2
+
+* Migrates from `ui.hash*` to `Object.hash*`.
+
 ## 4.1.1
 
 * Updates package description.