+++ packages/multicast_dns/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.2+1
+
+* Migrates from `ui.hash*` to `Object.hash*`.
+
 ## 0.3.2
 
 * Updates package description.